refactor(compare): extract C1+C2 to checks/n_preservation.py (Phase A 5/N)
Move check_c1_n_preservation (total participant count) and
check_c2_n_loss (per-variable valid N) into the new
hv_dataqc.compare.checks subpackage.

Creates checks/__init__.py for the subpackage. n_preservation.py imports
CheckResult and the fmt_n helper from _common.

Regression-check: COPDGene, ARIC, HCHS — MD + JSON all match. Tests:
79/79 pass.

Main file: 1733 → 1584 lines.
